requirements.txt
Flask==2.3.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
python-dotenv==1.0.1
passlib==1.7.4
pyotp==2.9.0
qrcode==7.4.2
cryptography==42.0.5
requests==2.31.0
